requirements.txt
# TruGrade Professional Platform Requirements
# Revolutionary Card Grading System Dependencies

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# AI/ML Core
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Computer Vision & Image Processing
albumentations>=1.3.0
imageio>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0

# UI Framework - PyQt6 (Primary GUI Framework)
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
PyQt6-sip>=13.6.0

# Legacy UI Support (being phased out)
customtkinter>=5.2.0

# Database & Caching
asyncpg>=0.29.0
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web & API
aiofiles>=23.2.0
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.31.0

# Async & Concurrency
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0
websockets>=12.0

# Data Processing
pandas>=2.1.0
polars>=0.19.0
pyarrow>=14.0.0

# Configuration & Environment
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

# Logging & Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0
psutil>=5.9.0

# Security
cryptography>=41.0.0
bcrypt>=4.1.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0

# Mobile PWA Support
qrcode[pil]>=7.4.0

# Revolutionary Features
# Photometric Stereo Analysis
open3d>=0.18.0
trimesh>=4.0.0

# Continuous Learning
wandb>=0.16.0
mlflow>=2.8.0

# Performance Optimization
numba>=0.58.0
cython>=3.0.0

# GPU Acceleration (Optional)
# torch-audio>=2.1.0  # Uncomment if audio processing needed
# torchaudio>=2.1.0   # Uncomment if audio processing needed

# Production Deployment
gunicorn>=21.2.0
supervisor>=4.2.0

# TensorZero Integration (Critical for AI/ML orchestration)
tensorzero>=1.0.0
transformers>=4.30.0

# Additional Dependencies
packaging>=21.0
typing-extensions>=4.5.0
